Запоздалое замечание по реестру
